added new files

+++ webfluxdemo/src/main/java/code/elif/webfluxdemo/service/input/CalculationInput.java
@@ -3,10 +3,12 @@ package code.elif.webfluxdemo.service.input;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CalculationInput {
 
-    private Integer number1;
-    private Integer number2;
+    private BigDecimal number1;
+    private BigDecimal number2;
 }
