+++ webfluxdemo/src/main/java/code/elif/webfluxdemo/service/output/CalculationOutput.java
@@ -5,10 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor(onConstructor_ = {@JsonCreator(mode = JsonCreator.Mode.PROPERTIES)})
 public class CalculationOutput extends BaseOutput{
 
-    private Integer result;
+    private BigDecimal result;
 }
